backend/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
python-dotenv=1.0.1